Zadanie24.Time. str20-22. konstructor

+++ lesson5/src/Zadanie24/Time.java
@@ -17,8 +17,8 @@ public class Time {
 
 
     public Time(int sec, int min, int hours){
-        this.hours= (int) ((sec + min*60+hours*3600)/3600);
-        this.min=(int)(((sec + min*60+hours*3600) - this.hours*3600)/60);
+        this.hours=  ((sec + min*60+hours*3600)/3600);
+        this.min=(((sec + min*60+hours*3600) - this.hours*3600)/60);
         this.sec=(sec + min*60L+hours*3600L) - this.hours* 3600L -this.min*60L;
     }
